CareDesk/requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[openai,deepgram,cartesia,silero,webrtc]>=0.0.99
